requirements.txt
pytest~=8.1.1
pytest-selenium == 4.1.0
termcolor
allure-pytest == 2.13.5
allure-python-commons
selenium~=4.19.0
webdriver-manager==4.0.1
urllib3~=2.2.1